backend/requirements.txt
Django==5.1.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
psycopg2-binary==2.9.10
gunicorn==22.0.0
pydantic==2.9.2
openai==1.52.2
httpx==0.27.2
chromadb==0.5.5
pypdf==5.0.1
docx2txt==0.8
python-dotenv==1.0.1
dj-database-url==2.1.0
whitenoise==6.6.0
